+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `Controller::fpga_info()` is now return vector of `FPGAInfo`
 * Change delay and duty offset API
 * Change `gain::Grouped` and `GainSequence` API
+* Remove `sequence::Circumference`
 
 # 1.9.3
 * fix minor bugs